(maint) Exclude files that cause YARD errors
Running `puppet strings server` generates documentation for all
modules in your modulepath, meaning that it also generates HTML
for strings itself. Since there are parts of strings that are not
meant to be documented with YARD, exclude the files that cause
error messages so that `puppet strings server` is not needlessly
noisy.

Additionally, make a few suggested changes to the README to make
it clearer and easier to read.