(maint) Require puppet version >= 3.7 for typing
Prior to this commit, the metadata for strings required a puppet
version of 3.6.0 or greater. However, now that we are deriving
parameter type information, we need 3.7.0 or greater because parameter
typing was not included in previous versions of the parser.

Update the metadata to reflect this change in version requirement.