(maint) Update YARD to version 0.9.3
Prior to this commit, puppet strings tests were failing against YARD
versions 0.9.0 and newer. This was due to 1) a change in the YARD
results output and 2) the upstream patch we submitted in PDOC-25
being released.

The upstream patch we submitted for that ticket added a regular
expression that we could override when matching the names of puppet
classes and defines since they begin with lowercase letters (unlike
Ruby). Until this was released with the new YARD gem, we were using
a different method of fixing this issue. Once the gem was out, the tests
started failing because the regex override doesn't propagate to the
tests.

In order to fix the failing tests, override the regex in the tests
as well.