(PDOC-63) Restore acceptance test to the spec directory.
Previously the acceptance test was implemented under the `spec` directory and
was moved out to a top level `acceptance` directory.

This broke the expected location for the node configuration files.

This commit puts the acceptance test back where it was previously.

Also fixes a failure in the test now that Markdown is the default markup
language used.