(PDOC-80) Remove runtime dependency on puppet
The runtime dependency on puppet presents problems when
puppet is not installed on the system as a gem, as is the
case in PE installations. Because the dependency will only be
satisfied by a gem, strings currently cannot be installed in
PE 3.8. This had to be worked around in newer PE versions with
a fake gemspec file.

Another issue is that installing strings with an older
version of puppet will force an upgrade to the newest version of
puppet via the gem.

This commit replaces the runtime dependency on puppet with a
requirement which will be present on the gem homepage on
rubygems.org. Requirements are meant to list external requirements
needed for the gem to work, which is correct for puppet as it can be
present via a different type of package.