(PDOC-160) Remove the timestamp from output
Prevent mass-changes to files by removing the date in the footer of each page.
Since the generation date isn't particularly useful and isn't tied to either
a git revision or a released version, this could be an easy way to address this
bug.  The files will of-course be regenerated but if there are no changes then
the content will be identical and git will not notice.