(MAINT) Bump travis puppet4 cell to Ruby 2.1.9
The README specifies a minimum Ruby version of 2.1.9, and we only
relaxed the gemspec requirement to 2.1.0 in PDOC-259 to support some
puppet_agent module testing. Skipping 2.1.0 for testing allows us to
bump our dev-only dependency on Beaker to a reasonably modern range.
Since there is no runtime dependency on Beaker, this shouldn't break
downstream users on Ruby 2.1.0.