(PDOC-226) Refactor Markdown testing
Previously the markdown unit testing used an a simple if statement to switch
between common and Puppet Plans.  However with the introduction of Puppet Types
we need another combination.  This commit refactors the Markdown testing use an
additive style to figure out test fixtures which makes adding conditional tests
easier in the future.