(PDOC-283) Fix namespaced symbols
Previously in Yard 0.9.19 dyna_symbols (symbols represented as strings e.g.
:'something') were converted to their unqualified symbol name e.g. :something.
However YARD treated this as a bug and in Yard 0.19.20 dynasymbols are now their
original source.  This commit modifies the base ruby handler to always convert
dyna_symbols to their plain name, as Puppet uses this for actual documentation
and Puppet Code.  For example the fully qualified function 'foo::bar' will
appear in the ruby function definition as `.. :'foo::bar'` however in the Puppet
documentation the name should appears as `foo::bar`.

This commit also adds tests for this scenario as there were no existing tests
for this.