(PDOC-295) Add 'enum' tag support for Enum data types
Enum data types would benefit from the ability to add extended
documentation on each option similarly to the '@option' tag
for hashes. The option type is not a good fit for Enum data
type parameters as it requires data types to be specified,
which doesn't make sense in the context of a Enum type.