(GH-225) Document functions in Puppet Datatypes
Previously the puppet datatype handler only documented attributes, not functions
for example [1].  This commit updates the data type handler to read in the
function information and then allow that to be used by puppet-strings in its
JSON, Markdown and HTML renderers.  This commit also adds tests for many
different scenarios and all three rendering types.

[1] https://github.com/puppetlabs/puppet/blob/3e03e734dda415272ec156b03ed9f021ba243228/lib/puppet/datatypes/error.rb#L13-L15