Catch parse failures in `PuppetStrings::Json` tests
Previously, the tests for `PuppetStrings::Json` (in json_spec.rb) parsed
a variety of code samples and then queried YARD to test the results. If
a sample failed to parse, an error might be outputted (see [issue
#307][#307]), but no failure was registered by the testing harness.

This adds simple checks to verify that the parses actually succeed.

Unfortunately, one of the samples currently fails to parse, so this
commit introduces a testing error. The subbsequent commit will fix it.

[#307]: https://github.com/puppetlabs/puppet-strings/issues/307