(#307) Fix plan parsing in JSON spec tests
Previously, the parsing code activated task support in Puppet when it
detected that it was parsing a file in a plans/ directory. This casued
the JSON spec test for plans to fail because it parsed a string, which
naturally had no file name to check.

This turns on task support (and thus enables parsing plans) whenever the
parser starts (assuming that task support is available).

This helps avoid surprises to future users who need to parse plans from
ruby code, such as in tests.