(#307) Fix plan parsing in JSON spec tests
Previously, parsing of Puppet plan code failed in json_spec.rb because
it didn’t provide a file name to the parser, which prevented the parser
from enabling task support in Puppet, which prevented the parser from
recognizing a plan.

This turns on task support in Puppet before trying to parse a plan in
json_spec.rb.