Use tilde heredocs for readability
This changes most heredocs to use the tilde syntax with variable
interpolation off. It indents the contents of the heredoc along with the
final symbol.

To maintain these changes, it enables `Layout/HeredocIndentation` and
`Layout/ClosingHeredocIndentation` in Rubocop.

The tilde syntax smartly strips the indentation, and allows for
different indention of the closing identifier:

```ruby
"test\n" == <<~'END'
    test
  END
```

This syntax was introduced in Ruby 2.3.0, so this will not pose
compatibility problems.

### Tests that should fail

Note that two tests couldn’t be updated to a more common heredoc format
to avoid failing.

It appears there is a problem with the way YARD parses parameters with
defaults with a heredoc ([YARD #779][]). There is a work-around for that
issue in `PuppetStrings::Yard::Handlers::Ruby::Base`, but it can’t work
in these cases because YARD doesn’t provide enough source to read the
whole heredoc content.

Given that these are tests of the Puppet 3 style functions, it’s
probably not worth pursuing this issue.

[YARD #779]: https://github.com/lsegal/yard/issues/779