(#223) Use code blocks as appropriate in Markdown
Sometimes data types, default values, and other code outputted in
Markdown documentation is multiline. When that’s the case, it’s cleaner
to use a code block (below) rather and inline code (`code`).

```
example code block
with two lines
```

This updates the Markdown template code to use code blocks in many
places if the code in question contains a newline.

Thanks to @crazymind1337 for the example of a multiline type alias!