(GH-332) Ensure PuppetStrings is loaded for tasks
Prior to this commit users would receive an Undefined Constant error
when attempting to generate a reference file via the provided Rake
tasks.

This was caused by the compact testing style for modules added in v3.0.

In this case the parent, PuppetStrings is not properly loaded before
Tasks. As a result the error above is caused.

This commit adds a new module declaration to tasks.rb that will ensure
that the PuppetStrings is properly loaded.