Rake task allows for a different commit message
Without this commit, you cannot inform the rake task to push gh-pages to
accept a different commit message. This is necessary in environment
where commit messages have to be in specific formats.

None of the system calls were checking the exit status of their
respective commands. Not stopping on error means that the rake take can
fail and you and your pipelines would not know.